+++ src/DrawUtils.hpp
@@ -45,7 +45,7 @@ void CST_RenderCopyRotate(CST_Renderer* dest, CST_Texture* src, CST_Rect* src_re
 
 
 void CST_SetDrawColor(CST_Renderer* renderer, CST_Color c);
-void CST_SetDrawColorRGBA(CST_Renderer* renderer, uint32_t r, uint32_t g, uint32_t b, uint32_t a);
+void CST_SetDrawColorRGBA(CST_Renderer* renderer, uint8_t r, uint8_t g, uint8_t b, uint8_t a);
 void CST_FillRect(CST_Renderer* renderer, CST_Rect* dimens);
 void CST_SetDrawBlend(CST_Renderer* renderer, bool enabled);
