+++ src/RootDisplay.hpp
@@ -21,7 +21,7 @@ public:
 
 	bool process(InputEvents* event);
 	void render(Element* parent);
-	void background(int r, int g, int b);
+	void background(uint8_t r, uint8_t g, uint8_t b);
   void initAndStartMusic();
 	void update();
